zcc-rpc-common/src/test/java/com/zcc/zccrpc/common/utils/ReflectionUtilsTest.java
package com.zcc.zccrpc.common.utils;

import org.junit.Test;

import java.lang.reflect.Method;

import static org.junit.Assert.*;

public class ReflectionUtilsTest {

    @Test
    public void newInstance() {
        TestClass testClass = ReflectionUtils.newInstance(TestClass.class);
        assertNotNull(testClass);
    }

    @Test
    public void getPublicMethods() {
        Method[] methods = ReflectionUtils.getPublicMethods(TestClass.class);
        assertEquals(1,methods.length);
    }

    @Test
    public void invoke() {
        Method[] methods = ReflectionUtils.getPublicMethods(TestClass.class);
        Method b = methods[0];

        TestClass testClass = ReflectionUtils.newInstance(TestClass.class);
        Object o = ReflectionUtils.invoke(testClass, b);
        assertEquals("b",o);


    }
}